cast logged image to float16

+++ src/trainer/base_trainer.py
@@ -496,7 +496,7 @@ class BaseTrainer:
         #print('logging image', batch["image"].shape, batch["image"])
         self.writer.add_image(
             image_name=mode,
-            image=F.to_pil_image(batch["image"].squeeze(dim=0).cpu()),
+            image=F.to_pil_image(batch["image"].squeeze(dim=0).to(torch.float16).cpu()),
         )
